added operations for doctor; added update operations

+++ backend/src/main/java/com/hoscrm/Patient/PatientService.java
@@ -1,7 +1,6 @@
 package com.hoscrm.Patient;
 
 import com.hoscrm.Appointment.AppointmentRepository;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Example;
 import org.springframework.stereotype.Service;
 
@@ -11,11 +10,9 @@ import java.util.Optional;
 @Service
 public class PatientService {
     public PatientRepository patientRepository;
-    private AppointmentRepository aRep; // try to inject in method
 
-    public PatientService(PatientRepository patientRepository, AppointmentRepository aRep){
+    public PatientService(PatientRepository patientRepository){
         this.patientRepository = patientRepository;
-        this.aRep = aRep;
     }
 
     public List<Patient> findPatients(String firstName, String lastName, Short age){
@@ -30,8 +27,21 @@ public class PatientService {
         return patientRepository.findById(id);
     }
 
-    public void deletePatientsByIds(List<Long> ids){
+    public boolean deletePatientsByIds(List<Long> ids){
+        if(!ids.stream().allMatch(i -> {
+            return patientRepository.existsById(i);
+        }))
+            return false;
         patientRepository.deleteAllById(ids);
+        return true;
+    }
+
+    public List<Patient> updatePatients(List<Patient> patients){
+        if(!patients.stream().allMatch(p -> {
+            return patientRepository.existsById(p.getId());
+        }))
+            return null;
+        return patientRepository.saveAll(patients);
     }
 
 }
